# api/openapi.yaml
openapi: 3.0.3
info:
  title: GreenPass 软件测试治理平台 API
  description: |
    测试质量管控平台后端契约（P1）。
    租户隔离：所有业务请求必须携带 `x-gp-team-id` 与可选 `x-gp-user-id` 头，经 RLS 过滤。
    审计哈希链 / 成本明细 / 门禁判定由 trusted 域统一写（口径单点）。
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /cases:
    get:
      summary: 用例列表
      operationId: listCases
      parameters:
        - { name: x-gp-team-id, in: header, required: true, schema: { type: integer } }
        - { name: target_id, in: query, schema: { type: integer } }
        - { name: kind, in: query, schema: { type: string } }
        - { name: code, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200': { description: 用例列表, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/TestCase' } } } } }
    post:
      summary: 创建用例（版本化首版）
      operationId: createCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ target_id, code, title ]
              properties:
                target_id: { type: integer }
                code: { type: string }
                title: { type: string }
                kind: { type: string }
                script: { description: 用例脚本/数据（执行产物来源） }
      responses:
        '201': { description: 已创建（含版本）, content: { application/json: { schema: { $ref: '#/components/schemas/TestCase' } } } }
  /cases/{id}/versions:
    post:
      summary: 迭代创建用例新版本
      operationId: createCaseVersion
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ script ]
              properties:
                script: { description: 用例脚本/数据 }
                source_repo_id: { type: integer }
                source_branch: { type: string }
      responses:
        '200': { description: 新版本已创建, content: { application/json: { schema: { $ref: '#/components/schemas/TestCase' } } } }
  /cases/{id}/rollback:
    post:
      summary: 用例按版本回退（避免错误分支生成）
      operationId: rollbackCase
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ version ]
              properties:
                version: { type: integer }
      responses:
        '200': { description: 已回退 }
  /cases/{id}/history:
    get:
      summary: 用例版本历史对比
      operationId: caseHistory
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 版本历史, content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } }
  /cases/{id}:
    delete:
      summary: 删除用例
      operationId: deleteCase
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 已删除, content: { application/json: { schema: { $ref: '#/components/schemas/StatusResponse' } } } }
  # ---------- 版本校验（execution，GP1-03） ----------
  /targets/{id}/env/runtime:
    post:
      summary: 登记测试环境运行版本（env_runtime）
      operationId: registerRuntime
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ env, version ]
              properties:
                env: { type: string }
                version: { type: string }
      responses:
        '200': { description: 已登记, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
  /targets/{id}/version-check:
    post:
      summary: 发起版本校验（match/mismatch/unknown）
      operationId: checkVersion
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ env, target_version ]
              properties:
                env: { type: string }
                target_version: { type: string }
      responses:
        '200':
          description: 校验结果（mismatch 阻断信号）
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { type: string, enum: [ match, mismatch, unknown ] }
  /targets/{id}/version-checks:
    get:
      summary: 最近版本校验记录
      operationId: recentChecks
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        '200': { description: 记录列表, content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } } }
  # ---------- 运行编排（execution，GP1-04） ----------
  /runs:
    post:
      summary: 创建测试运行（勾选用例空=全量）
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ target_id, scenario_id, target_version ]
              properties:
                target_id: { type: integer }
                scenario_id: { type: integer }
                env: { type: string }
                target_version: { type: string }
                target_branch: { type: string }
                selected_case_ids: { type: array, items: { type: integer }, description: 空=全量 }
      responses:
        '201': { description: 已创建（queued）, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
  /runs/{id}:
    get:
      summary: 运行详情（状态机）
      operationId: getRun
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: 运行聚合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
  /runs/{id}/case-results:
    get:
      summary: 运行内用例结果
      operationId: runCaseResults
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 用例结果列表, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/CaseResult' } } } } }
  /runs/{id}/version-check:
    post:
      summary: 触发版本校验阶段
      operationId: startVersionCheck
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: mismatch 阻断, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
  /runs/{id}/execute:
    post:
      summary: 执行运行（重跑 attempt 递增）
      operationId: executeRun
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 已启动, content: { application/json: { schema: { type: object, required: [ run, results ], properties: { run: { $ref: '#/components/schemas/Run' }, results: { type: array, items: { $ref: '#/components/schemas/CaseResult' } } } } } } }
  /runs/{id}/pause:
    post:
      summary: 暂停
      operationId: pauseRun
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 已暂停, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
  /runs/{id}/resume:
    post:
      summary: 恢复
      operationId: resumeRun
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 已恢复, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
  # ---------- 门禁（trusted，GP1-05） ----------
...
components:
  schemas:
    StatusResponse:
      type: object
      required: [ status ]
      properties:
        status: { type: string }
...
    TestCase:
      type: object
      required: [ id, team_id, target_id, code, title, kind, current_version, status ]
      properties:
        id: { type: integer, format: int64 }
        team_id: { type: integer, format: int64 }
        target_id: { type: integer, format: int64 }
        code: { type: string }
        title: { type: string }
        kind: { type: string }
        current_version: { type: integer }
        status: { type: string }
    Run:
      type: object
      required: [ id, team_id, scenario_id, target_id, env, target_version, target_branch, env_version, run_mode, state, selected_cases ]
      properties:
        id: { type: integer, format: int64 }
        team_id: { type: integer, format: int64 }
        scenario_id: { type: integer, format: int64 }
        target_id: { type: integer, format: int64 }
        env: { type: string }
        target_version: { type: string }
        target_branch: { type: string }
        env_version: { type: string }
        env_check_id: { type: integer, format: int64, nullable: true }
        run_mode: { type: string }
        state: { type: string, enum: [ queued, version_check, scheduled, running, paused, collect, gate, report, done, failed ] }
        selected_cases: { type: array, items: { type: integer, format: int64 } }
        started_at: { type: string, format: date-time, nullable: true }
        ended_at: { type: string, format: date-time, nullable: true }
    CaseResult:
      type: object
      required: [ id, team_id, run_id, case_id, case_version, status, result_text, attempt_seq ]
      properties:
        id: { type: integer, format: int64 }
        team_id: { type: integer, format: int64 }
        run_id: { type: integer, format: int64 }
        case_id: { type: integer, format: int64 }
        case_version: { type: integer }
        status: { type: string, enum: [ pass, fail, blocked, skipped, retried ] }
        result_text: { type: string }
        evidence: { type: object, additionalProperties: true, nullable: true }
        attempt_seq: { type: integer }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time, nullable: true }
...
  securitySchemes:
    teamHeader:
      type: apiKey
      in: header
      name: x-gp-team-id
security:
  - teamHeader: []